requirements.txt
# 51job爬虫系统依赖包

# 核心爬虫框架
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
cryptography>=41.0.0  # MySQL连接加密支持

# 配置管理
PyYAML>=6.0

# 任务调度
schedule>=1.2.0
APScheduler>=3.10.0

# 网络请求增强
requests-toolbelt>=1.0.0
urllib3>=2.0.0
chardet>=5.0.0

# 数据验证
cerberus>=1.3.0
validators>=0.20.0

# 日志处理
coloredlogs>=15.0
loguru>=0.7.0

# 工具库
tqdm>=4.65.0  # 进度条
click>=8.1.0  # 命令行工具
rich>=13.0.0  # 富文本输出

# 时间处理
python-dateutil>=2.8.0
pytz>=2023.3

# 文件处理
openpyxl>=3.1.0  # Excel支持

# 网络工具
fake-useragent>=1.4.0

# 开发工具（可选）
ipython>=8.0.0
jupyter>=1.0.0

# 测试工具（可选）
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# 代码质量（可选）
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# 性能监控（可选）
memory-profiler>=0.61.0
psutil>=5.9.0

# 异步支持（可选）
aiohttp>=3.8.0
aiofiles>=23.0.0

# 图像处理（可选，用于验证码）
Pillow>=10.0.0
opencv-python>=4.8.0

# 机器学习（可选，用于数据分析）
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 缓存（可选）
redis>=4.6.0

# 消息队列（可选）
celery>=5.3.0
kombu>=5.3.0

# 监控（可选）
prometheus-client>=0.17.0

# 部署（可选）
gunicorn>=21.0.0
supervisor>=4.2.0